release v5.0x1.1
4d6e761 aiq: Update version to v5.0x1.1
3c8842c hwi: Sync kernel driver header to v2.1.0
b7d94ec j2s: parser tool: support both arm and arm64
6d266f8 SmartIr: update smartIr sample
185aa77 Adrc and adehaze: support camgroup res
690b863 CCM: fix bug in ccm_querry_info api func
3ca507f AE: support set FpsMax in Auto-Fps Mode
6b63ebc AE: get fps for reference
3f9f9d2 supplement commit 9e19894
eb5f269 add uapi support isp3x
...


Former-commit-id: 4721cabc25c372b78b4a2326b2b96c4048bf33c0

+++ rkaiq/algos/aeis/remap_backend.cpp
@@ -201,10 +201,10 @@ FecMeshBuffer* FecRemapBackend::GetPendingHwResult() {
                       [&](const std::unique_ptr<FecMeshBuffer>& p) {
                           if (p->State[0] == fec_hw_mesh_used_by_hardware ||
                               p->State[0] == fec_mesh_skipped) {
-                              if (p->FrameId != -1 && p->FrameId <= last_result_id_) {
+                              if (p->FrameId != (uint32_t)(-1) && p->FrameId <= last_result_id_) {
                                   LOGW_AEIS("Get pending result id %u PASSED !!!", p->FrameId);
                                   p->State[0] = fec_mesh_available;
-                              } else if (last_result_id_ != -1 && p->FrameId - last_result_id_ > 1) {
+                              } else if (last_result_id_ != (uint32_t)(-1) && p->FrameId - last_result_id_ > 1) {
                                   LOGV_AEIS("pending result id %u in FUTURE!!!", p->FrameId);
                               } else {
                                   if (min_id >= p->FrameId) {
